SBI LIFE INSURANCE
columns/attributes details:
---------------------------------------------------------------------------------
age: age of the policyholder 

sex: gender of the policyholder. It has 2 categories as follows;
  Female=0
  Male=1 

bmi: Body mass index, providing an understanding of the body, weights that are relatively high or low relative to height, objective index of body weight (kg / m ^ 2) using the ratio of height to weight, ideally 18.5 to 25 

steps: average walking steps per day of a policyholder 

children: number of children/dependents of the policyholder 

smoker: smoking state of the policyholder.  it has 2 categories as follows;
   non_smokeer=0,  
   smoker=1)

region: the residential area of the policyholder in the US.  it has 4 categories as follows
     Northeast=0, 
     Northwest=1, 
     Southeast=2
     Southwest=3
 
charges: individual medical costs billed by health insurance 

insuranceclaim:  it has 2 categories as follows 
        yes=1,
        no=0


Task:
--------------------------------------------------------------------------------
Predict the Insurance Claim
--------------------------------------------------------------------------------
Problem Type: Classification